+++ failfast/src/test/kotlin/failfast/problematic/EmptyRootContextTest.kt
@@ -2,7 +2,6 @@ package failfast.problematic
 
 import failfast.describe
 
-// this is the place for test suites that are somehow problematic but should be handled gracefully
 
 object EmptyRootContextTest {
     val context = describe("empty root context") {