+++ failfast/src/test/kotlin/failfast/problematic/package-info.java
@@ -1,0 +1,3 @@
+package failfast.problematic;
+
+// this is the place for test suites that are somehow problematic but should be handled gracefully