+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Mocking. Failfast now has its own mocking framework. It's very simple and can only mock interfaces, but it's much
-  faster than mockk, and has a nice api. (for now look into MockTest.kt for more info.)
+  faster than mockk, and has a nice api. (Look in the kdoc of Mock.kt or MockTest.kt for more info)
 - Allow declaring multiple root contexts in one class file. This is especially useful to create root contexts
   dynamically, for example to run the same test cases on multiple databases.